Refactored the whole project
The HTML was unreadable in the previous version. Changed the whole of the HTML/CSS so classses have way more readable names. Unified repeated classes. Fixed a lot of responsiveness issues (Remade all of the media queries). Tried a completely different approach with the usage of background images on the second section of the project.