Fixed a bug
There was a bug with color of the curved background and the background-color of the section. Since the background color was converted to hsl, the color didn't exactly match with the one in the background images. It was also specified to the background images to be 2px talles than the padding, to ensure they always sit inside the container. It was also added a # to the action of the form, to specify that this is merely a dummy form.